SWATCH-3025: Create swatch-common-export module
There is no porting functionality here, but only move the logic into a common library that then will be reused by Quarkus and Spring Boot services.
