+++ build.gradle
@@ -34,6 +34,7 @@ allprojects {
 
 dependencies {
     implementation project(":swatch-common-clock")
+    implementation project(":swatch-common-export")
     implementation project(':swatch-core')
     implementation project(":api")
     implementation project(":clients:rh-marketplace-client")
@@ -55,10 +56,6 @@ dependencies {
     implementation "com.fasterxml.jackson.module:jackson-module-jakarta-xmlbind-annotations"
     // used to connect to umb
     implementation "org.springframework.boot:spring-boot-starter-activemq"
-    // used to parse events generated by console.redhat.com (Export Service)
-    implementation libraries["event-schemas"]
-    // used by the export feature to write CSV files
-    implementation "com.fasterxml.jackson.dataformat:jackson-dataformat-csv"
     implementation "org.liquibase:liquibase-core"
     implementation "org.postgresql:postgresql"
     implementation libraries["guava"]
