+++ settings.gradle
@@ -54,6 +54,7 @@ include ':swatch-common-splunk'
 include ':swatch-common-platform-spring-boot'
 include ':swatch-common-platform-quarkus'
 include ':swatch-common-clock'
+include ':swatch-common-export'
 include ':swatch-metrics'
 include ':swatch-common-trace-response'
 include ':swatch-metrics-hbi'
