+++ src/main/java/org/candlepin/subscriptions/ApplicationConfiguration.java
@@ -30,7 +30,6 @@ import com.fasterxml.jackson.dataformat.csv.CsvMapper;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.fasterxml.jackson.module.jakarta.xmlbind.JakartaXmlBindAnnotationModule;
-import com.redhat.cloud.event.parser.ConsoleCloudEventParser;
 import io.micrometer.core.aop.TimedAspect;
 import io.micrometer.core.instrument.MeterRegistry;
 import jakarta.validation.Validator;
@@ -47,7 +46,6 @@ import org.candlepin.subscriptions.subscription.SubscriptionServiceConfiguration
 import org.candlepin.subscriptions.tally.TallyWorkerConfiguration;
 import org.candlepin.subscriptions.util.LiquibaseUpdateOnlyConfiguration;
 import org.candlepin.subscriptions.util.UtilConfiguration;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.autoconfigure.info.ConditionalOnEnabledInfoContributor;
 import org.springframework.boot.actuate.autoconfigure.info.InfoContributorFallback;
 import org.springframework.context.ApplicationContext;
@@ -171,10 +169,5 @@ public class ApplicationConfiguration implements WebMvcConfigurer {
   @ConditionalOnEnabledInfoContributor(value = "certs", fallback = InfoContributorFallback.DISABLE)
   public CertInfoContributor certInfoContributor(ApplicationContext context) {
     return new CertInfoContributor(context);
-  }
-
-  @Bean
-  public ConsoleCloudEventParser cloudEventParser(@Autowired ObjectMapper objectMapper) {
-    return new ConsoleCloudEventParser(objectMapper);
   }
 }
