+++ src/main/java/org/candlepin/subscriptions/capacity/CapacityIngressConfiguration.java
@@ -21,7 +21,7 @@
 package org.candlepin.subscriptions.capacity;
 
 import org.candlepin.subscriptions.db.RhsmSubscriptionsDataSourceConfiguration;
-import org.candlepin.subscriptions.export.ExportSubscriptionConfiguration;
+import org.candlepin.subscriptions.export.ExportConfiguration;
 import org.candlepin.subscriptions.resteasy.ResteasyConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigurationExcludeFilter;
 import org.springframework.boot.context.TypeExcludeFilter;
@@ -43,7 +43,7 @@ import org.springframework.scheduling.annotation.EnableAsync;
 @Import({
   ResteasyConfiguration.class,
   RhsmSubscriptionsDataSourceConfiguration.class,
-  ExportSubscriptionConfiguration.class
+  ExportConfiguration.class
 })
 @ComponentScan(
     basePackages = {
