+++ src/main/java/org/candlepin/subscriptions/subscription/export/SubscriptionDataExporterService.java
@@ -33,6 +33,10 @@ import static org.candlepin.subscriptions.resource.ResourceUtils.ANY;
 
 import com.redhat.swatch.configuration.registry.MetricId;
 import com.redhat.swatch.configuration.registry.ProductId;
+import com.redhat.swatch.export.DataExporterService;
+import com.redhat.swatch.export.DataMapperService;
+import com.redhat.swatch.export.ExportServiceException;
+import com.redhat.swatch.export.ExportServiceRequest;
 import jakarta.ws.rs.core.Response;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -52,10 +56,6 @@ import org.candlepin.subscriptions.db.model.ServiceLevel;
 import org.candlepin.subscriptions.db.model.SubscriptionCapacityView;
 import org.candlepin.subscriptions.db.model.SubscriptionCapacityViewMetric;
 import org.candlepin.subscriptions.db.model.Usage;
-import org.candlepin.subscriptions.exception.ExportServiceException;
-import org.candlepin.subscriptions.export.DataExporterService;
-import org.candlepin.subscriptions.export.DataMapperService;
-import org.candlepin.subscriptions.export.ExportServiceRequest;
 import org.candlepin.subscriptions.json.SubscriptionsExportJsonMeasurement;
 // NOTE(khowell): this couples our export implementation to the v1 REST API
 import org.candlepin.subscriptions.util.ApiModelMapperV1;
