+++ src/main/java/org/candlepin/subscriptions/tally/export/InstancesCsvDataMapperService.java
@@ -23,6 +23,8 @@ package org.candlepin.subscriptions.tally.export;
 import com.redhat.swatch.configuration.registry.MetricId;
 import com.redhat.swatch.configuration.registry.Variant;
 import com.redhat.swatch.configuration.util.MetricIdUtils;
+import com.redhat.swatch.export.DataMapperService;
+import com.redhat.swatch.export.ExportServiceRequest;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
@@ -34,8 +36,6 @@ import org.candlepin.subscriptions.db.HostRepository;
 import org.candlepin.subscriptions.db.model.HardwareMeasurementType;
 import org.candlepin.subscriptions.db.model.Host;
 import org.candlepin.subscriptions.db.model.TallyInstanceView;
-import org.candlepin.subscriptions.export.DataMapperService;
-import org.candlepin.subscriptions.export.ExportServiceRequest;
 import org.candlepin.subscriptions.json.InstancesExportCsvItem;
 import org.candlepin.subscriptions.util.ApiModelMapperV1;
 import org.springframework.data.domain.Page;
