+++ src/main/java/org/candlepin/subscriptions/tally/export/InstancesDataExporterService.java
@@ -26,6 +26,10 @@ import static org.candlepin.subscriptions.resource.api.v1.InstancesResource.getH
 
 import com.redhat.swatch.configuration.registry.MetricId;
 import com.redhat.swatch.configuration.registry.ProductId;
+import com.redhat.swatch.export.DataExporterService;
+import com.redhat.swatch.export.DataMapperService;
+import com.redhat.swatch.export.ExportServiceException;
+import com.redhat.swatch.export.ExportServiceRequest;
 import jakarta.ws.rs.core.Response;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
@@ -45,10 +49,6 @@ import org.candlepin.subscriptions.db.model.ServiceLevel;
 import org.candlepin.subscriptions.db.model.TallyInstanceView;
 import org.candlepin.subscriptions.db.model.TallyInstancesDbReportCriteria;
 import org.candlepin.subscriptions.db.model.Usage;
-import org.candlepin.subscriptions.exception.ExportServiceException;
-import org.candlepin.subscriptions.export.DataExporterService;
-import org.candlepin.subscriptions.export.DataMapperService;
-import org.candlepin.subscriptions.export.ExportServiceRequest;
 // NOTE(khowell): this couples our export implementation to the v1 REST API
 import org.candlepin.subscriptions.utilization.api.v1.model.ReportCategory;
 import org.springframework.context.annotation.Profile;
