+++ swatch-common-export/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+    id 'swatch.java-library-conventions'
+}
+
+dependencies {
+    // used to parse events generated by console.redhat.com (Export Service)
+    api libraries["event-schemas"]
+    // used by the export feature to write CSV files
+    api "com.fasterxml.jackson.dataformat:jackson-dataformat-csv"
+}
+
+description = 'Shared Export classes for swatch services'
