+++ swatch-common-export/src/main/java/com/redhat/swatch/export/JsonExportFileWriter.java
@@ -18,23 +18,21 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.subscriptions.export;
+package com.redhat.swatch.export;
+
+import static com.redhat.swatch.export.ExportRequestHandler.INTERNAL_ERROR;
 
 import com.fasterxml.jackson.core.JsonEncoding;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.ws.rs.core.Response;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.stream.Stream;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.candlepin.subscriptions.exception.ExportServiceException;
-import org.springframework.stereotype.Service;
 
 @Slf4j
-@Service
 @AllArgsConstructor
 public class JsonExportFileWriter implements ExportFileWriter {
 
@@ -66,8 +64,7 @@ public class JsonExportFileWriter implements ExportFileWriter {
       jGenerator.writeEndObject();
     } catch (IOException e) {
       log.error("Error writing the Json payload for request {}", request.getExportRequestUUID(), e);
-      throw new ExportServiceException(
-          Response.Status.INTERNAL_SERVER_ERROR.getStatusCode(), "Error writing the Json payload");
+      throw new ExportServiceException(INTERNAL_ERROR, "Error writing the Json payload");
     }
   }
 
@@ -77,7 +74,6 @@ public class JsonExportFileWriter implements ExportFileWriter {
         item,
         request.getExportRequestUUID(),
         e);
-    throw new ExportServiceException(
-        Response.Status.INTERNAL_SERVER_ERROR.getStatusCode(), "Error writing the payload");
+    throw new ExportServiceException(INTERNAL_ERROR, "Error writing the payload");
   }
 }
